backend/requirements.txt
fastapi
uvicorn[standard]
nexla-sdk
python-dotenv
